Fix #1904: support `$hybrid` for Lexical INSERT
