+++ src/test/java/io/stargate/sgv2/jsonapi/api/v1/InsertInCollectionIntegrationTest.java
@@ -29,7 +29,6 @@ import java.util.regex.Pattern;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.bson.types.ObjectId;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.api.ClassOrderer;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Nested;
@@ -1002,138 +1001,6 @@ public class InsertInCollectionIntegrationTest extends AbstractCollectionIntegra
       givenHeadersPostJsonThenOkNoErrors("{ \"insertOne\": { \"document\": %s }}".formatted(doc))
           .body("$", responseIsWriteSuccess())
           .body("status.insertedIds[0]", is("doc-with-path-overlap"));
-    }
-  }
-
-  @Nested
-  @Order(5)
-  class InsertOneWithLexical {
-    static final String DOC_WITH_LEXICAL =
-        """
-                {
-                  "_id": "lexical1",
-                  "username": "user-lexical",
-                  "extra": 123,
-                  "$lexical": "monkeys and bananas"
-                }
-                """;
-
-    // NOTE: Relies on default settings allowing insertion of "$lexical" content
-    @Test
-    @Order(1)
-    public void insertDocWithLexicalOk() {
-      // Must have Lexical functionality available
-      Assumptions.assumeTrue(isLexicalAvailableForDB());
-
-      givenHeadersPostJsonThenOkNoErrors(
-                  """
-                  {
-                    "insertOne": {
-                      "document": %s
-                    }
-                  }
-                  """
-                  .formatted(DOC_WITH_LEXICAL))
-          .body("$", responseIsWriteSuccess())
-          .body("status.insertedIds[0]", is("lexical1"));
-
-      givenHeadersPostJsonThenOkNoErrors(
-              """
-                      {
-                        "find": {
-                          "filter" : {"_id" : "lexical1"}
-                        }
-                      }
-                      """)
-          .body("$", responseIsFindSuccess())
-          // NOTE: "$lexical" is not included in the response by default, ensure
-          .body(
-              "data.documents[0]",
-              jsonEquals(
-                  """
-                                  {
-                                      "_id": "lexical1",
-                                      "username": "user-lexical",
-                                      "extra": 123
-                                  }
-                                  """));
-
-      // But can explicitly include: either via "include-it-all"
-      givenHeadersPostJsonThenOkNoErrors(
-              """
-                      {
-                        "find": {
-                          "filter" : {"_id" : "lexical1"},
-                          "projection": { "*": 1 }
-                        }
-                      }
-                      """)
-          .body("$", responseIsFindSuccess())
-          .body("data.documents[0]", jsonEquals(DOC_WITH_LEXICAL));
-
-      // Or just the "$lexical" field (plus always _id)
-      givenHeadersPostJsonThenOkNoErrors(
-              """
-                      {
-                        "find": {
-                          "filter" : {"_id" : "lexical1"},
-                          "projection": {
-                            "$lexical": 1,
-                            "extra": 1
-                          }
-                        }
-                      }
-                      """)
-          .body("$", responseIsFindSuccess())
-          .body(
-              "data.documents[0]",
-              jsonEquals(
-                  """
-                                            {
-                                                "_id": "lexical1",
-                                                "extra": 123,
-                                                "$lexical": "monkeys and bananas"
-                                            }
-                                            """));
-    }
-
-    @Test
-    @Order(2)
-    public void failInsertDocWithLexicalIfNotEnabled() {
-      final String COLLECTION_WITHOUT_LEXICAL =
-          "coll_insert_no_lexical_" + RandomStringUtils.randomNumeric(16);
-      createComplexCollection(
-              """
-                      {
-                        "name": "%s",
-                        "options" : {
-                          "lexical": {
-                            "enabled": false
-                          }
-                        }
-                      }
-                      """
-              .formatted(COLLECTION_WITHOUT_LEXICAL));
-
-      givenHeadersAndJson(
-                  """
-                  {
-                    "insertOne": {
-                      "document": %s
-                    }
-                  }
-                  """
-                  .formatted(DOC_WITH_LEXICAL))
-          .when()
-          .post(CollectionResource.BASE_PATH, keyspaceName, COLLECTION_WITHOUT_LEXICAL)
-          .then()
-          .body("$", responseIsWritePartialSuccess())
-          .body("errors", hasSize(1))
-          .body("errors[0].errorCode", is("LEXICAL_NOT_ENABLED_FOR_COLLECTION"))
-          .body("errors[0].message", containsString("Lexical search is not enabled"));
-
-      // And finally, drop the Collection after use
-      deleteCollection(COLLECTION_WITHOUT_LEXICAL);
     }
   }
